# Conflicts:
#	Source/CMakeLists.txt
#	Source/Filters/ConvolutionFreq.h
#	Source/Filters/ITD.h

+++ Source/AudioPackage/PackageEncoder.h
@@ -38,7 +38,7 @@ public:
 
 private:
 
-    unsigned bufferSize = 0;
+    uint64_t bufferSize = 0;
 
     struct FileInfo
     {
